Add git submodule for proto files + split CMakeLists.txt
Proto files, previously in the src/model folder, have been moved to
their own git repository. This commit adds this repository to the
project direcotry tree as a git submodule.

Also, the CMakeLists.txt is split in two. The executable target is split
also in two: the executable and exe library (so that it can be later
unit-tested).