Change the name and return value of Sqlite3Functions::SelectTags.
The function's name is changed to SelectRows() and is now used for a
more general purpose of executing a SELECT query on any table. Instead
of a vector of model::Tag's it now returns the collected string values.
The conversion of the "raw" values is the responsibility of the caller.