Fix bug where insertion into Tagged table failed when saving new Memo.
When inserting a new Memo that had tags assigned to it, after it has
been inserted into the db, an incorrect memo id was used to insert new
rows into the Tagged table. The id used was the one coming from the
'request' Memo that didn't have its id initialised to any value (the
value was random). Insertions to Tagged table failed due to foreign-key
contstraints in SQL.
This commit fixes the bug by using the correct memo id, i.e. the one
assigned to the newly inserted memo.